+++ src/compas_cem/equilibrium/force_numpy.py
@@ -6,9 +6,6 @@ from compas_cem.diagrams import FormDiagram
 __all__ = ["static_equilibrium_numpy"]
 
 
-SMALL_VALUE = 1e-3
-
-
 def static_equilibrium_numpy(topology, tmax=100, eta=1e-6, verbose=False, callback=None):
     """
     Generate a form diagram in static equilibrium using numpy.
@@ -149,35 +146,27 @@ def equilibrium_state_numpy(topology, tmax=100, eta=1e-6, verbose=False, callbac
                 # override length if a plane exists
                 if plane is not None:
                     # get length from line plane intersection
-                    p_length = trail_edge_length_from_plane_numpy(pos, rvec, plane)
+                    plength = trail_length_from_plane_intersection_numpy(pos, rvec, plane)
 
                     # check that returned length is not null
-                    if p_length is None:
-                        # print("Warning! No intersection found between edge {} and plane {}".format(edge, plane))
-                        # print("Falling back to input length: {}".format(length))
-                        p_length = length
-                    # print out warning if zero length.
-                    # elif p_length < SMALL_VALUE:
-                    #     msg = "Warning! Length for edge after {} intersection is {}. Check the plane."
-                    #     print(msg.format(edge, p_length))
-
-                    # correct length sign based on internal force state
-                    dot = p_length * length
-                    if dot < 0:
-                        length = p_length * -1.0
-                    else:
-                        length = p_length
+                    if plength:
+                        length = plength
 
                 # compute trail force
-                trail_force = length_vector(rvec)  # always positive
+                trail_force = length_vector_numpy(rvec)  # always positive
+
+                # NOTE: to avoid NaNs, do not normalize residual vector if it is zero length
+                nrvec = rvec / trail_force
+                if np.isnan(length_vector_numpy(nrvec)):
+                    nrvec = rvec
 
                 # store next node position
-                next_pos = pos + length * rvec / trail_force
+                next_pos = pos + length * nrvec
                 node_xyz[next_node] = next_pos
 
                 # correct trail force sign based on trail signed length
-                # TODO: autograd.np does not support derivatives on copysign
-                if trail_force * length < 0:  #
+                # NOTE: autograd.np does not support derivatives on copysign
+                if trail_force * length < 0.0:
                     trail_force = trail_force * -1.0
 
                 # store trail force
@@ -229,8 +218,7 @@ def form_update(form, node_xyz, trail_forces, reaction_forces):
     """
     # assign nodes' coordinates
     for node, xyz in node_xyz.items():
-        form.node_xyz(node, xyz)
-        # form.node_attributes(key=node, names=["x", "y", "z"], values=xyz)
+        form.node_attributes(key=node, names=["x", "y", "z"], values=xyz)
 
     # assign forces on trail edges
     for edge, tforce in trail_forces.items():
@@ -245,7 +233,7 @@ def form_update(form, node_xyz, trail_forces, reaction_forces):
     for edge in form.edges():
         # length = form.edge_length(u, v)
         u, v = edge
-        length = length_vector(node_xyz[u] - node_xyz[v])
+        length = length_vector_numpy(node_xyz[u] - node_xyz[v])
         force = form.edge_attribute(key=edge, name="force")
         length = np.copysign(length, force)
         form.edge_attribute(key=(u, v), name="length", value=length)
@@ -423,66 +411,55 @@ def vector_two_nodes(a, b, normalize=False):
     vector = a - b
     if not normalize:
         return vector
-    return normalize_vector(vector)
+    return normalize_vector_numpy(vector)
 
 
-def normalize_vector(vector):
+def normalize_vector_numpy(vector):
     """
     Hand-made vector normalization.
     """
-    return vector / length_vector(vector)
+    return vector / length_vector_numpy(vector)
 
 
-def length_vector(vector):
+def length_vector_numpy(vector):
     """
     Calculates the norm of a vector.
     """
     return np.linalg.norm(vector)
 
 
-def trail_edge_length_from_plane_numpy(pos, direction, plane):
+def trail_length_from_plane_intersection_numpy(point, vector, plane, tol=1e-6):
     """
-    Calculates the length of a trail edge from a line-plane intersection.
+    Calculates the signed length of a trail edge from a vector-plane intersection.
 
     Parameters
     ----------
-    pos : ``np.array``
-        The XYZ coordinates of the starting node of an edge.
-    direction : ``np.array``
-        The XYZ coordinates of the direction the edge points to.
-    plane : ``tuple`` of ``np.array``
-        The origin node and the normal vector of a plane.
+    point : ``list`` of ``float``
+        The XYZ coordinates of the base position of the vector.
+    direction : ``list`` of ``float``
+        The XYZ coordinates of the vector.
+    plane : ``Plane``
+        A COMPAS plane defined by a base point and a normal vector.
+    tol : ``float``, optional
+        A tolerance to check if vector and the plane normal are parallel
+        Defaults to ``1e-6``.
 
     Returns
     -------
-    length : ``float``
+    length : ``float``, ``None``
         The distance between ``pos`` and the resulting line-plane intersection.
         If not intersection is found, it returns ``None``.
-
-    Notes
-    -----
-    A line will be created by adding ``pos`` and ``direction``.
-    The line functions as a line ray, not as finite-length segment.
     """
-    a = pos
-    b = a + direction
+    origin, normal = plane
+    cos_nv = np.dot(normal, normalize_vector_numpy(vector))
 
-    # intersection line plane
-    o, n = plane
+    if np.abs(cos_nv) < tol:
+        return
 
-    ab = b - a
-    cosa = np.dot(n, ab)
+    oa = origin - point
+    cos_noa = np.dot(normal, oa)
 
-    if np.abs(cosa) <= SMALL_VALUE:
-        # raise ValueError("np abs cosa {} <= {}, pos:{}, vec:{}, normal:{}".format(np.abs(cosa), SMALL_VALUE, pos, ab, n))
-        return None
-
-    oa = a - o
-    ratio = - np.dot(n, oa) / cosa
-    ab = ab * ratio
-    pos_plane = a + ab
-
-    return length_vector(pos - pos_plane)
+    return cos_noa / cos_nv
 
 
 if __name__ == "__main__":