+++ src/compas_cem/ghpython/components/CompasCem_ConstrainedFormFinding/code.py
@@ -8,10 +8,11 @@ from compas.rpc import Proxy
 
 
 class ConstrainedFormFindingComponent(component):
-    def RunScript(self, solve, topology, constraints, parameters, algorithm, iters_max, eps, tmax, eta):
+    def RunScript(self, solve, topology, constraints, parameters, algorithm, iters_max, eps, kappa, tmax, eta):
         algorithm = algorithm or "SLSQP"
         iters_max = iters_max or 100
         eps = eps or 1e-6
+        kappa = kappa or 1e-8
         tmax = tmax or 100
         eta = eta or 1e-6
 
@@ -39,6 +40,7 @@ class ConstrainedFormFindingComponent(component):
                                    algorithm=algorithm,
                                    iters=iters_max,
                                    eps=eps,
+                                   kappa=kappa,
                                    tmax=tmax,
                                    eta=eta)