+++ src/compas_cem/ghpython/components/CompasCem_ParameterNodeOriginY/code.py
@@ -8,7 +8,7 @@ from compas_cem.optimization import OriginNodeYParameter
 
 class OriginNodeYParameterComponent(component):
     def RunScript(self, node_key, bound_low, bound_up):
-        if node_key is None or bound_low is None or bound_up is None:
+        if node_key is None:
             return
 
         return OriginNodeYParameter(node_key, bound_low, bound_up)