+++ src/compas_cem/optimization/optimizer.py
@@ -1,10 +1,10 @@
 from time import time
 
+from functools import partial
+
 import autograd.numpy as np
 
 from autograd import grad as agrad
-
-from functools import partial
 
 from compas_cem.data import Data
 
@@ -16,6 +16,9 @@ from compas_cem.optimization import grad_finite_differences
 from compas_cem.optimization import objective_function_numpy
 from compas_cem.optimization import nlopt_solver
 from compas_cem.optimization import nlopt_status
+
+from compas_cem.optimization.parameters import EdgeParameter
+from compas_cem.optimization.parameters import NodeParameter
 
 from nlopt import RoundoffLimited
 
@@ -34,8 +37,8 @@ class Optimizer(Data):
     def __init__(self, **kwargs):
         super(Optimizer, self).__init__(**kwargs)
 
-        self.parameters = {}
-        self.constraints = {}
+        self.parameters = dict()
+        self.constraints = dict()
 
         self.x_opt = None
         self.time_opt = None
@@ -43,6 +46,9 @@ class Optimizer(Data):
         self.evals = None
         self.gradient_norm = None
         self.status = None
+
+        self._ckey = -1
+        self._pkey = -1
 
 # ------------------------------------------------------------------------------
 # Counters
@@ -61,58 +67,42 @@ class Optimizer(Data):
         return len(self.constraints)
 
 # ------------------------------------------------------------------------------
-# Mappings
-# ------------------------------------------------------------------------------
-
-    def index_parameter(self):
-        """
-        A dictionary that maps indices to parameter keys.
-        """
-        return {idx: key for idx, key in enumerate(self.parameters.keys())}
-
-    def parameter_index(self):
-        """
-        A dictionary that maps parameter keys to indices.
-        """
-        return {key: idx for key, idx in self.index_parameter()}
-
-# ------------------------------------------------------------------------------
-# Additions
+# Parameters
 # ------------------------------------------------------------------------------
 
     def add_parameter(self, parameter):
         """
         Adds a parameter to the optimization problem.
         """
-        key = (parameter.key(), parameter.attr_name())
-        self.parameters[key] = parameter
+        self._pkey += 1
+        self.parameters[self._pkey] = parameter
+
+    def remove_parameter(self, pkey):
+        """
+        Removes an optimization parameter.
+        """
+        if pkey not in self.parameters:
+            raise KeyError("Parameter not found at object key: {}".format(pkey))
+        del self.parameters[pkey]
+
+# ------------------------------------------------------------------------------
+# Constraints
+# ------------------------------------------------------------------------------
 
     def add_constraint(self, constraint):
         """
         Adds a goal constraint.
         """
-        key = constraint.key()
-        self.constraints[key] = constraint
+        self._ckey += 1
+        self.constraints[self._ckey] = constraint
 
-# ------------------------------------------------------------------------------
-# Removals
-# ------------------------------------------------------------------------------
-
-    def remove_parameter(self, key):
+    def remove_constraint(self, ckey):
         """
-        Removes an optimization parameter.
+        Removes a constraint from the optimizer.
         """
-        if key not in self.parameters:
-            raise KeyError("Parameter not found at object key: {}".format(key))
-        del self.parameters[key]
-
-    def remove_constraint(self, key):
-        """
-        Removes a goal constraint from the optimizer.
-        """
-        if key not in self.constraints:
-            raise KeyError("Constraints not found on object key: {}".format(key))
-        del self.constraints[key]
+        if ckey not in self.constraints:
+            raise KeyError("Constraints not found on object key: {}".format(ckey))
+        del self.constraints[ckey]
 
 # ------------------------------------------------------------------------------
 # Objective Function
@@ -137,15 +127,13 @@ class Optimizer(Data):
         x_func = partial(self._optimize_form, topology=topology, tmax=tmax, eta=eta)
         return partial(grad_f, x_func=x_func, step_size=step_size)
 
-# ------------------------------------------------------------------------------
+# ---------------------- --------------------------------------------------------
 # Solver
 # ------------------------------------------------------------------------------
 
-    def solve(self, topology, algorithm="SLSQP", grad="AD", step_size=1e-6, iters=100, eps=1e-6, tmax=100, eta=1e-6, verbose=False):
+    def solve(self, topology, algorithm="SLSQP", grad="AD", step_size=1e-6, iters=100, eps=1e-6, kappa=1e-8, tmax=100, eta=1e-6, verbose=False):
         """
         Solve a constrained form-finding problem using gradient-based optimization.
-
-        The gradient of the objective function computed using automatic differentiation.
 
         Parameters
         ----------
@@ -168,18 +156,19 @@ class Optimizer(Data):
             The method to compute the gradient of the objective function.
             The currently available methods are:
 
-            - AD: Automatic differentiation.
-            - FD: Finite differences.
+            - AD: Automatic differentiation
+            - FD: Finite differences
 
             Defaults to "AD".
         iters : ``int``, optional
             The maximum number of iterations to run the optimization algorithm for.
             Defaults to ``100``.
         eps : ``float``, optional
-            The numerical convergence threshold for the optimization algorithm.
-            If value is set to ``None``, this parameter is ignored and the
-            optimization algorithm will run until ``iters`` is exhausted.
-            Defaults to ``None``.
+            The convergence threshold for the output value of the objective function.
+            Defaults to ``1e-6``.
+        kappa : ``float``, optional
+            The convergence threshold for the norm of the gradient of the objective function.
+            Defaults to ``1e-8``.
         step_size : ``float``, optional
             The step size to calculate the gradient of the objective function via finite differences.
             It becomes active only if ``grad="FD"``. It is otherwise ignored by this function.
@@ -245,7 +234,7 @@ class Optimizer(Data):
                             "bounds_up": bounds_up,
                             "iters": iters,
                             "eps": eps,
-                            "ftol": None}
+                            "ftol": kappa}
 
         # assemble optimization solver
         solver = nlopt_solver(**hyper_parameters)
@@ -262,10 +251,9 @@ class Optimizer(Data):
             print("Results may still be useful though!")
             x_opt = self.optimization_parameters(topology)
         except RuntimeError:
-             print("Optimization failed for reasons I do not understand yet...")
-             print(f"Optimization total runtime: {round(time() - start, 4)} seconds")
-             return static_equilibrium(topology)
-
+            print("Optimization failed for reasons I do not understand yet...")
+            print(f"Optimization total runtime: {round(time() - start, 4)} seconds")
+            return static_equilibrium(topology)
 
         # fetch last optimum value of loss function
         time_opt = time() - start
@@ -300,22 +288,20 @@ class Optimizer(Data):
 # Optimization parameters
 # ------------------------------------------------------------------------------
 
-    def optimization_parameters(self, form):
+    def optimization_parameters(self, topology):
         """
         Creates optimization paremeters array.
         Only one entry in the array per constraint.
         Takes care of keeping the ordering.
         """
-        # TODO: This can become an (n, 3) array to acount for root nodes variables?
-        x = np.zeros(self.number_of_parameters())
+        parameters = np.zeros(self.number_of_parameters())
 
-        for index, ckey in self.index_parameter().items():
-            constraint = self.parameters[ckey]
-            x[index] = constraint.start_value(form)
+        for pkey, parameter in self.parameters.items():
+            parameters[pkey] = parameter.start_value(topology)
 
-        return x
+        return parameters
 
-    def optimization_bounds(self, form):
+    def optimization_bounds(self, topology):
         """
         Creates optimization bounds array.
         Only one entry in the array per constraint.
@@ -323,10 +309,9 @@ class Optimizer(Data):
         bounds_low = np.zeros(self.number_of_parameters())
         bounds_up = np.zeros(self.number_of_parameters())
 
-        for index, ckey in self.index_parameter().items():
-            parameter = self.parameters[ckey]
-            bounds_low[index] = parameter.bound_low(form)
-            bounds_up[index] = parameter.bound_up(form)
+        for pkey, parameter in self.parameters.items():
+            bounds_low[pkey] = parameter.bound_low(topology)
+            bounds_up[pkey] = parameter.bound_up(topology)
 
         return bounds_low, bounds_up
 
@@ -334,55 +319,26 @@ class Optimizer(Data):
 # Updates
 # ------------------------------------------------------------------------------
 
-    def _update_topology_origin_nodes(self, x, topology):
+    def _update_parameters(self, topology, parameters):
         """
+        Update the defined design parameters in a topology diagram.
         """
-        map_xyz_index = {"x": 0, "y": 1, "z": 2}
+        for pkey, parameter in self.parameters.items():
 
-        for index, ckey in self.index_parameter().items():
+            value = parameters[pkey]
+            name = parameter.attr_name()
+            key = parameter.key()
 
-            node = self.parameters[ckey].key()
-
-            # TODO: weak check, needs to be handled differently
-            if not isinstance(node, int):
-                continue
-
-            # TODO: this check should happen upon assembly, not during calculation?
-            if not topology.is_node_origin(node):
-                msg = "{} is not a root node. Assigned constraint is invalid!"
-                raise ValueError(msg.format(node))
-
-            # TODO: refactor to handle xyz more transparently
-            xyz = topology.node_xyz(key=node)
-            parameter = self.parameters[ckey]
-            j = map_xyz_index[parameter.attr_name()]
-            xyz[j] = x[index]
-
-            # self.form.node_xyz(key=node, xyz=xyz)  # form.node_xyz(node, y=x[])?
-            topology.node_attributes(key=node, names="xyz", values=xyz)
-
-    def _update_topology_edges(self, x, topology):
-        """
-        """
-        for index, ckey in self.index_parameter().items():
-
-            edge = self.parameters[ckey].key()
-
-            # TODO: weak check, needs to be handled differently
-            if isinstance(edge, int):
-                continue
-
-            if topology.is_trail_edge(edge):
-                name = "length"
-            elif topology.is_deviation_edge(edge):
-                name = "force"
-
-            value = x[index]
-
-            topology.edge_attribute(key=edge, name=name, value=value)
+            if isinstance(parameter, NodeParameter):
+                topology.node_attribute(key=key, name=name, value=value)
+            elif isinstance(parameter, EdgeParameter):
+                topology.edge_attribute(key=key, name=name, value=value)
+            else:
+                msg = "Parameter {} is neither a node nor an edge parameter! {}"
+                raise TypeError(msg.format(type(parameter)))
 
 # ------------------------------------------------------------------------------
-# Optimization
+# Penalty function
 # ------------------------------------------------------------------------------
 
     def _calculate_penalty(self, eq_state):
@@ -394,11 +350,14 @@ class Optimizer(Data):
 
         return penalty
 
+# ------------------------------------------------------------------------------
+# Optimization
+# ------------------------------------------------------------------------------
+
     def _optimize_form(self, parameters, topology, tmax, eta):
         """
         """
-        self._update_topology_origin_nodes(parameters, topology)
-        self._update_topology_edges(parameters, topology)
+        self._update_parameters(topology, parameters)
 
         eq_state = equilibrium_state_numpy(topology, tmax, eta)