+++ src/compas_cem/optimization/parameters/__init__.py
@@ -30,5 +30,10 @@ from __future__ import print_function
 from .parameter import *  # noqa F403
 from .edge import *  # noqa F403
 from .node import *  # noqa F403
+from .load import *  # noqa F403
+from .origin import *  # noqa F403
+from .trail import *  # noqa F403
+from .deviation import *  # noqa F403
+
 
 __all__ = [name for name in dir() if not name.startswith('_')]