+++ CHANGELOG.rst
@@ -11,23 +11,59 @@ Unreleased
 
 **Added**
 
+**Changed**
+
+**Fixed**
+
+**Deprecated**
+
+**Removed**
+
+0.6.1
+----------
+
+**Added**
+
+**Changed**
+
+**Fixed**
+
+- Fixed mising `kappa` argument in `optimization.solve_proxy`
+
+**Deprecated**
+
+**Removed**
+
+0.6.0
+----------
+
+**Added**
+
 - `ConstrainedFormFinding` GH component outputs optimized `TopologyDiagram`.
 - `compas_cem.plotters.TopologyArtist` has node size and edge width default parameters.
 - `compas_cem.plotters.FormArtist` has node size and edge width default parameters.
 - Added support for gradient computation via finite differences.
 - Added support for the `VAR` optimization algorithm from NLOpt.
+- Added gradient convergence treshold `kappa` to the signature of `Optimizer.solve`.
+- Added `kappa` as input to the `ConstrainedFormFinding` GH component.
+- Implemented `Diagram.is_edge_supported`
 
 **Changed**
 
 - Changed GH components folder prefix from CEM to CompasCem.
 - Renamed `compas_cem.optimization.Optimizer.solve_nlopt` to `solve`.
 - Refactored `compas_cem.optimization.Parameter`.
+- Refactored calculation of trail edge length via plane intersection in `static_equilibrium`.
+- Refactored calculation of trail edge length via plane intersection in `static_equilibrium_numpy`.
 
 **Fixed**
 
 - Fixed bug in `compas_cem.equilibrium.equilibrium_state` where `residual` calculation was undefined when `tmax=1`.
 
 **Deprecated**
+
+- Deprecated `Optimizer.solve_nlopt` and `.solve_nlopt_proxy`.
+
 
 **Removed**